style: Rename function to be more generic and apply for future tickets
