+++ src/utils/index.ts
@@ -2,4 +2,4 @@ export { default as localizeDate } from './localizeDate';
 export { default as localizeDateTime } from './localizeDateTime';
 export { default as extractVideoEmbedFromStory } from './extractVideoEmbedFromStory';
 export { default as isServerSide } from './serverSide';
-export { default as fixTrailingSlash } from './fixTrailingSlash';
+export { default as formatURL } from './formatURL';
